requirements.txt
selenium >= 4.5.0
chromedriver-py == 106.0.5249.21
async-generator >= 1.10
attrs >= 22.1.0
certifi>=2022.9.24
chromedriver-py>=106.0.5249.21
exceptiongroup >= 1.0.0
gunicorn >= 20.1.0
h11 >= 0.14.0
idna >= 3.4
MouseInfo >= 0.1.3
outcome >= 1.2.0
pip >= 22.2.2
python-dotenv >= 0.21.0

selenium >= 4.5.0
setuptools >= 65.4.0
sniffio >= 1.3.0
sortedcontainers >= 2.4.0
trio >= 0.22.0
trio-websocket >= 0.9.2
urllib3 >= 1.26.12
wheel >= 0.37.1
wsproto >= 1.2.0